requirements.txt
numpy==2.2.0
opencv_python==4.10.0.84
Pillow==10.1.0
Pillow==11.0.0
streamlit==1.41.1
tensorflow==2.18.0
tensorflow_intel==2.18.0
